tests for query editor
