+++ public/app/plugins/datasource/grafana-azure-monitor-datasource/components/QueryEditor.test.tsx
@@ -1,54 +1,22 @@
 import React from 'react';
-import { render, screen, waitFor, fireEvent } from '@testing-library/react';
+import { render, screen, waitFor } from '@testing-library/react';
+import userEvent from '@testing-library/user-event';
 
 import QueryEditor from './QueryEditor';
-import Datasource from '../datasource';
 
 import mockQuery from '../__mocks__/query';
+import createMockDatasource from '../__mocks__/datasource';
 import { AzureQueryType } from '../types';
-
-type DeepPartial<T> = {
-  [P in keyof T]?: DeepPartial<T[P]>;
-};
-
-function makeMockDatasource() {
-  // We make this a partial so we get _some_ kind of time safety when making this, rather than
-  // having it be any or casted immediately to Datasource
-  const _mockDatasource: DeepPartial<Datasource> = {
-    azureMonitorDatasource: {
-      isConfigured() {
-        return true;
-      },
-      getSubscriptions: jest.fn().mockResolvedValueOnce([]),
-    },
-
-    getResourceGroups: jest.fn().mockResolvedValueOnce([]),
-    getMetricDefinitions: jest.fn().mockResolvedValueOnce([]),
-    getResourceNames: jest.fn().mockResolvedValueOnce([]),
-    getMetricNamespaces: jest.fn().mockResolvedValueOnce([]),
-    getMetricNames: jest.fn().mockResolvedValueOnce([]),
-    getMetricMetadata: jest.fn().mockResolvedValueOnce({
-      primaryAggType: 'average',
-      supportedAggTypes: [],
-      supportedTimeGrains: [],
-      dimensions: [],
-    }),
-  };
-
-  const mockDatasource = _mockDatasource as Datasource;
-
-  return mockDatasource;
-}
 
 describe('Azure Monitor QueryEditor', () => {
   it('renders the Metrics query editor when the query type is Metrics', async () => {
-    const mockDatasource = makeMockDatasource();
+    const mockDatasource = createMockDatasource();
     render(<QueryEditor query={mockQuery} datasource={mockDatasource} onChange={() => {}} />);
     await waitFor(() => expect(screen.getByTestId('azure-monitor-metrics-query-editor')).toBeInTheDocument());
   });
 
   it("does not render the Metrics query editor when the query type isn't Metrics", async () => {
-    const mockDatasource = makeMockDatasource();
+    const mockDatasource = createMockDatasource();
     const logsMockQuery = {
       ...mockQuery,
       queryType: AzureQueryType.LogAnalytics,
@@ -58,15 +26,18 @@ describe('Azure Monitor QueryEditor', () => {
   });
 
   it('changes the query type when selected', async () => {
-    const mockDatasource = makeMockDatasource();
+    const mockDatasource = createMockDatasource();
     const onChange = jest.fn();
     render(<QueryEditor query={mockQuery} datasource={mockDatasource} onChange={onChange} />);
-    // TODO: select "Logs" from the Query Type dropdown
+    await waitFor(() => expect(screen.getByTestId('azure-monitor-query-editor')).toBeInTheDocument());
+
+    // Click the dropdown, then select "Logs"
+    userEvent.click(screen.getByText('Metrics'));
+    userEvent.click(screen.getByText('Logs'));
 
     expect(onChange).toHaveBeenCalledWith({
       ...mockQuery,
       queryType: AzureQueryType.LogAnalytics,
     });
-    // await waitFor(() => expect(screen.getByTestId('azure-monitor-query-editor')).toBeInTheDocument());
   });
 });
